・On the client side, create a service layer by adding a new "Services" folder. ・Utilize our new ProductService within our ProductList view, rather than directly using Http to access the API. ・From this point forward, services will handle accessing the API and similar tasks, while views will rely on these services to retrieve the necessary data.
※ Take note of how our ProductService class includes a static property, "Products". This property represents the list of products that our service handles.
As a result of this design, we can directly use the "Products" property whenever we need to refer to our list of products, without having to declare a new variable and assign this list to it every time.
This simplifies our code by allowing direct access to the product list through the ProductService, and eliminates the need for additional variable declarations for storing the list in each usage scenario.
In the GetProducts method, the ProductService handles retrieving the product list from the API and directly assigns this data to the "Products" property. This way, the updated product list is readily accessible through the ProductService anytime after the GetProducts method has been called.